src/backend/requirements.txt
aiohttp==3.8.2
aiosignal==1.2.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
ecdsa==0.17.0
fastapi==0.100.0
frozenlist==1.3.1
greenlet==3.0.3
h11==0.13.0
httplib2==0.20.4
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
multidict==5.2.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
pg8000==1.24.1
platformdirs==2.5.1
protobuf==3.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
rsa==4.8
scramp==1.4.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
yarl==1.8.1
urllib3==1.26.9
uvicorn==0.17.6
yarl==1.8.1
